Replace 'react-popper' with '@floating-ui/react-dom'
The thread list rendered in the demo dashboards was not showing the
imageModal and tooltips properly. I didn't manage to get to the bottom
of the reason why but noticed that our tooltips use a different
popper than that that was wrapping the thread list in the demo app.


Test Plan: Ran the app locally and can see it works.

Reviewers: jozef-mokry, flooey

Reviewed By: jozef-mokry, flooey

Pull Request: https://github.com/getcord/monorepo/pull/2450

monorepo-commit: 43bc962fffc3cdf3f03e3837436216eeb3ed5cd0