Enable hiding header row on playground apps (#2530)
For user research, we want a version of the playground apps that
doesn't have a header so that users can't accidentally or
intentionally click away from the area they're supposed to be doing
their tasks.  Respect a top-level CSS class of .demo-no-header and
adjust the layout to remove the header.



Test Plan:

Set the class on the body element on each of the three playground
apps, the header disappeared.

monorepo-commit: b3e11363010574819136ea4cb43cf9cb255f8d3c