Update the dashboard to use presence.useLocationData
The hook that it is using right now (useCordPresentUsers) is deprecated.
In fact, the dashboard crashes in prod right now because it assumes that
it is getting back the data in the shape that presence.useLocationData
returns, while in fact it is still getting back data in the shape
useCordPresentUsers is returning (or rather was returning before
useCordPresentUsers was just a re-export of useLocationData in our
latest npm package).


Reviewers: jwatzman, flooey, maylynn-ng

Reviewed By: jwatzman, flooey, maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/4969

monorepo-commit: 553af10d8645be1eb95faa6ef5edb430dbbd4c2d