Use cord-pin and cord-thread on Dashboard sample app
This PR replaces our FloatingThreads + Annotations API setup with
cord-pin and cord-thread. The new approach leverages metadata on threads
to store information about where on the chart or table the comment was
made.

The line chart on the dashboard uses Pin and Thread. Pins are positioned
absolutely over the chart and their positions are only recomputed on
chart's `redraw` event. Thus avoiding the performance cost of having to
update the pins on every page scroll.

The table grid uses only Thread (and PresenceFacepile) that is
positioned `absolute` with respect to the table grid container. The
thread's position is updated automatically on table scroll by the
`floating-ui` library. The extra work required for the table grid was
detecting when the row with the open thread is scrolled out of view.

Because this code directly works with Threads (as opposed to
FloatingThreads component), it can easily choose both `metadata` and
`threadID` for each thread. For chart comments I derive threadIDs from
the point information (x, y, seriesId, chartId), whilst for table grid I
derive the threadIDs from the cell information (rowId, column id, grid
id). This ensures that there is always at most 1 comment per chart point
/ grid cell.

This approach makes it easy to choose which parts of a page are
annotatable. In this app, for the chart I had to add a point click-event
handler. For table grid I had to add `onCellClicked` handler. Compared
to Annotations API this is both a more direct way of doing it, and also
does not have the performance cost of having to check on every mouse
move whether the cursor is over annotatable area or not. Another related
example of what can be done with the new approach I change the mouse
cursor when you hover over table cell or chart point.


Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/5082

monorepo-commit: 04ace294bc376edcaaa4f6fa115e41d6f02f52a1