Disallow threadList and notificationList open at the same time
It does not look good on the dashboard sample app when the
notificationList opens on top of the thread list. The other way around
is not an issue because notificationList closes itself automatically
when you click outside.

I am also removing the goToComment prop because it can be just read from
threadsContext directly.


Reviewers: flooey, mhtien

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/5195

monorepo-commit: be0c48210ec2a1230ea66c9916bfbec6bb8607da