Fix attachment previews in Dashboard sample app
The Thread and ThreadList use a "position: fixed" modal overlay for
display attachment previews. Fixed positioning is with respect to the
closest containing block (initially viewport). However, using CSS
transform starts a new containing block which mean that the overlay was
not covering the viewport anymore. For now I rewrote the CSS to not use
"transform", but in the long run we should factor out the image
attachment modal to live outside of components so that it can
be attached to `<body>` directly.


Test Plan: Image attachments work on both the table and chart

Reviewers: #launch, Netceer, flooey

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/5209

monorepo-commit: 95f074bbb850c25ce9796f2f120d57eaf0933bd5