Set up type-aware linting
To get this to work in our monorepo, every TS file needs to be part of
some tsconfig.json. Notably the stuff in `sample-apps/_common` was not,
so copy a tsconfig from one of the other apps over there.


Test Plan:
CI build is happy, which includes a TS run, a lint run, and a build.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/5268

monorepo-commit: 7ebcd02e1c42a278771d9e2374d985b4f812441f