Fix an issue where clicking the video would not play/pause the video
This is just a strange bug, not sure why, and I couldn't find an explanation with google.

But, if you have `controlsList='nofullscreen'` as an attribute on a `video` element, the default click behavior for the video stops working.  If we remove that, and allow for full screen, the default behavior comes back and a click will play/pause the video.


Test Plan:
Used the demo apps, saw that I could now play/pause the video as expected

Reviewers: mhtien, flooey

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/5690

monorepo-commit: 3f3672b45f5c8b87008d1e518ea2d43952bbe3f4