Document demo app: Add selection comments (2/6)
Adding the selection comment functionality (from GDocs):
- When you select text, a "Comment" button appears.
- Clicking that button creates a thread, where the metadata stored is used to then retrieve the exact selected text
- The ThreadsContext is the same as the Dashboard. Just changed the Metadata we store and a few minor things.

This PR would have been easier, except that I had to implement GDocs logic: if you add two or more threads on the same line, they don't overlap. They are magically positioned.
Since this is demo code, LMK if it's followable.

The CSS is still scuffed. The idea is to fix that in the next PR.

Feel free to patch this PR and play with it!

Test Plan:
Test app locally.


Reviewers: flooey, Netceer

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/5955

monorepo-commit: 178c259d3e21cfb9c1fb53a12ea707464c9f2991