Add panning and zooming into canvas capabilities (4/n)
This PR addings panning and zooming capabilities to the canvas
with the pins adjusting their positions. It also ensures that the pins
pointer events are set to none when the user is panning to prevent the
pan from stopping if you hovered over a pin.

There is more to do with whena user decies to pan when they are on the
pin or thread but I've left that as something to do later.

Test Plan:
Tested locally and can see the pins move when you pan canvas.


Reviewers: Netceer, AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6146

monorepo-commit: 49d1ddbf046d47d9e12f69daea3043d13505038c