Upgrade TypeScript to 5.1.6 (#6219)
We're several versions behind, so move us forward.  Nothing huge in
the new versions, but a couple quality-of-life updates.

https://devblogs.microsoft.com/typescript/announcing-typescript-5-0/
https://devblogs.microsoft.com/typescript/announcing-typescript-5-1/

As part, switch from `^x.y.z` to `~x.y.z` versions for TypeScript,
since TypeScript uses the minor version for feature releases that can
have backwards incompatibilities.

This shouldn't affect any customers, since this is only a dev
dependency of ours, so it doesn't transfer to users of these packages.

This also required upgrading a few dependencies, but none of them were
major or interesting upgrades.

Test Plan:


For opensource/sdk-js:

* `npm run tsc-packages` - no errors

For opensource/sample-apps:

* In each app, `npm run build` - no errors

monorepo-commit: 11797f590828dc65b10d7882790c1b11b3fb1743