Document Demo App:Fix commenting on Firefox (#6283)
Summary:
You couldn't comment by clicking on the button's text on firefox.
I assume it is because of the click deselecting the text before the
commenting callback is executed.

Not allowing the button's text to be selected, avoids the selected text
to be unselected, and fixes it. :shrug:


Test Plan:
On firefox/chrome, add comment by clicking on the button icon and the
button text.

Reviewers:

monorepo-commit: 39ec153e9170f7a062fb82e80fbed70752648753