Video demo: Disable fullscreen
We originally disabled it, but then turned it back on because somehow this property disabled the "click to play" video functionality.

However, now we changed the UX: clicking on the video never plays it, it leaves a comment, so disabling fullscreen again.

Test Plan:
Click on video, check I can leave a comment.
Check fullscreen mode is disabled.

Reviewers: flooey, dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/6303

monorepo-commit: 6530005590fc940e627c5c401a22001bbdab9b64