Document demo: Fix threads flicker
When adding a comment, the bottom-most thread would flicker.

This was because the visibility was also controlled by `threadsPositions[threadIdx]`.

When adding a new comment, `sortedThreads` gets update before `threadsPositions`, as it takes a couple of re-renders to compute the latter.
For the above reason, I'm also removing the check threadsReady.size === sortedThreads.length we had in place to avoid extra re-renders (turns out it didn't really save us)

Test Plan:
Add a `debugger;` in the render, check the render cycles and the various pieces of state we compute.
Check that inserting a comment anywhere in the doc now (before, in the middle or after other comments) nicely animates, moving other threads around, without flickering.


Reviewers: flooey, Netceer

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/6307

monorepo-commit: 64e5ce7e8794382a6a3fd0cab7b535b548faca79