Make pins load on video control when video is paused
This PR ensures that the pins on the controls load up if the video is
initially paused. I haven't removed the auto play yet as the video is a
live demo and might be odd without the other parts that we will add
(the new video itself... an open thread etc).


Test Plan: Locally comment out the autoPlay and can see the pins!

Reviewers: flooey, AlbertQM, Netceer

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6349

monorepo-commit: db4188f46e35d0ab033639787b6052463d9f5c31