Ad new video to video demo docs and adapt playground token messages
This adds a new video, and loads different ones (ratio wise)
dependent on if it's on the website or the docs.

Changed the time of the comments on the playground token to be below 52
seconds (the duration of the video). Also changes the number so it is in
sync with the video. Subject to change.


Reviewers: anothertompetty, flooey, Netceer

Reviewed By: anothertompetty, Netceer

Pull Request: https://github.com/getcord/monorepo/pull/6369

monorepo-commit: eab4fdad36fcc84eb0d58e6ac1eca7e69a5a1286