Video demo: Fix onVideoClick logic
On initial load, the tooltip was wrong. Fixed that by using the `isPlaying` state, which should be up-to-date even if we change `autoplay` again in the future.

And then, there was an extra `.pause()`, removing that.

Test Plan:
Clicking on video now always toggles it, and the tooltip is correct.

Reviewers: flooey, mhtien, Netceer

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/6409

monorepo-commit: 9880684c5589443629c002a3b1d82715bade4b10