Video demo: position thread based on Pin's x/y
So far, we only checked which half (left or right) of the video the pin was, and moved the thread accordingly.

In this PR, we'll also start checking which half (top or bottom) and move the thread accordingly.


Test Plan:
Create a thread in every corner of the video, check it behaves as expected.

Reviewers: flooey, Netceer, mhtien

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/6442

monorepo-commit: 3e6ff1bccb0761b013c9d569f9fea9d7f90c92d0