Fix tooltip in Highchart dashboard demo
For some reason when we view the dashboard demo app on the cord website or docs
the tooltip cuts off some padding when you hover over a point on the graph.

No idea why it looks fine on the opensource repo but not on anything else.

I think it's related to perhaps the image (the comment icon) not loading
in time for the high charts svg to compute the size of it so I've set the
height and width of image.


Test Plan: Seems to work

Reviewers: Netceer, flooey

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/6457

monorepo-commit: e1556e8deb252e9fee509adc521016fc5c2a6192