Show different date range in website dashboard demo
We want to show a different date range on the website demo than the docs
demo.

This PR allows us to pass in a different date range, and we do for the website.
Also includes hiding the date range selector buttons if only one date range
existings. Also amended the data so we show 2012 to 2022 only.


Reviewers: AlbertQM, Netceer, flooey

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/6465

monorepo-commit: 0c9bb9db063bcd235258000c288512a293786d76