Document video: allow updating presence when thread open
When a thread is open, we render an underlay, so that we can close the thread if a user clicks on it (i.e. click outside the thread? --> close it)
However, we still want to update the presence on the document.
To do so, I'm moving the underlay to be rendered before the sheet, and adding `isolation: isolate` to create a stacking context.

Test Plan:
Add a black background with 0.6 opacity to the underlay. Check to see what is now above it.
When a thread is open, select soe text and add another thread (previously not possible). Looks ok!

<img width="1116" alt="image" src="https://github.com/getcord/monorepo/assets/33761650/c62d5e8d-f107-4e39-bb50-8644033f59d3">



Reviewers: mhtien, Netceer, flooey

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/6492

monorepo-commit: 7f340f76e14cdaf51bba863d0f739f47fa7de62a