Make autogenerated thread opened on dashboard demo app
This isn't ideal, but I've added an extra field in metadata
to check if this is an autogenerated message. If it is we set the
open thread id to this one.

Removed the reply from the dashboard autogenerated message as it was
too long for the dashboard demo on the webiste. As the content is
TBC we can fix this at a later date. Also changed the max-height of
the thread to 300px.

This currently only works because we have one autogenerated thread.

Reviewers: flooey, AlbertQM, Netceer

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6484

monorepo-commit: d54c388fd8bb7e076b609cafe1ded1e4a460164e