Document demo: Recompute threads when finishing typing animation
In the playground, we add two dummy comments.
These are added _before_ the content is inside the sheet (because we do the typing animation).
This meant `sortedThread` didn't find where to position the threads. and they ended up in the wrong order.

Test Plan:
Build app, visit docs locally, check the threads are now in the right order


Reviewers: flooey, Netceer, mhtien

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/6567

monorepo-commit: 3c7c5d225d27bb6e81fb4764becaa149598f843b