New document demo app - improving message hover state
The hover state on the message of a thread on the new document app highlighted
up to the left edge of the text, which didn't look great.

I noticed that in Google Docs, the avatar is aligned with the composer, and the message
content is shifted inwards by a few pixels - so I did that in our app as well. I also made sure
the right side of the message content is aligned with the composer.

![image](https://github.com/getcord/monorepo/assets/102963329/31d5a784-fa80-4e2a-85f0-09cfc761f97c)

Test Plan:
Before

<img width="221" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/cbeb708b-f28e-4f66-ae36-633a584e62c1">

After
<img width="243" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/54794ae5-a117-4bf6-9f06-5c162d127d8d">




Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6750

monorepo-commit: 83b2917f20d7664322b3e89dc318ab80e3a9fed4