demo-apps: video player - open thread doesn't block notification launcher
The position of the thread underlay "covered" the notification launcher, so if a thread was open you would have to click the launcher twice to open the notifications dropdown.

This PR moves the div so that you only have to click it once.

### To note:
In this implementation, clicking the notification launcher doesn't close the open thread - not sure if that's an issue? More a question of UX I guess.


#### Right side - current version. Double click on notification to open the drop down
#### Left side - PR version. Only click once on the notification

https://github.com/getcord/monorepo/assets/62358728/7a1beb36-9777-4144-aa74-7efc644278b6




Test Plan:
Open a thread & try to open the notification launcher. Should only have to click it once.

Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6777

monorepo-commit: 8beb30f1aebacbdff0b8b077e35a66069e238c84