New document demo: Fixing editing state in unfocused thread
When triggering message editing for a message and the thread is not open in the document demo, we reach a weird state where we can't see the composer. Looking at Google Docs, a thread can still be edited when unfocused. So this PR unhides the editing state composer when a thread isn't open.

Also, when the thread opens and is in editing state, we lose focus to the main composer of the thread. That's very frustrating! This case has also been handled by only setting autofocus to the thread when we are not editing.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/3813f9e5-68f0-4595-b750-399f89d5bcd9




Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6778

monorepo-commit: 2e9e1320b94e61daa3e70dac1f881c942d8c41e1