New document demo: Adding ThreadedComments
When deleting or resolving a thread, there is no way to retrieve it. To be able to also view these threads, I have added a ThreadedComments component. This PR adds the component and the relevant styles - with the actual GoogleDocs as a reference. I have mostly re-used the ThreadedCommentsLauncher component from the dashboard demo.

The thread layout is slightly different in the Thread component and the ThreadedComments component, so I had to modify message styles for ThreadedComments only. This also means that I have modified the specificity of some selectors for the message component, to only apply to the cord-thread component. For example the hover state of the message content. Some major differences in ThreadedComments, apart from the aforementioned hover state are: the message content is shifted inwards (not starting under the avatar), there are no reactions at all, and all replies are shifted to the left (kind of like an inline thread) - in the same column as the first message. The same applies for all buttons and their text.

Editing, deleted and unread states have been considered. I've also tried to order the ThreadedComments styles based on looking at the component from top to bottom thread > buttons > message etc etc. I thought it made more sense while reading them and no specificity ties need to be broken by order.

For now, resolved threads are interleaved, but we may add tabs, or a dropdown (like GDocs) in the future.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/80c1ac99-2468-4e05-8fbe-b08e1474dfe6


https://github.com/getcord/monorepo/assets/102963329/c772cc9f-4fea-465b-8ec3-1864cf8cf05d




Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6785

monorepo-commit: 46f52106e1827c70f6dcc524533a4d2f5e11fe08