New document demo: closing ThreadedComments on clicking outside the container
It is not right behavior to be interacting both with the threads on the document and
with the ones in ThreadedComments at the same time. I have changed the behavior to close the
ThreadedComments container when clicking outside it.

I have reused the underlay styles for the threads but added one more class to the ThreadedComments,
to give it the correct z-index. That is, so when we click on an open thread, we still close ThreadedComments.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/d611bb10-d972-4d45-9263-7307063dab4c





Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6811

monorepo-commit: 784e70b1d783bbf75d9b0739d9aff19705d122e1