New document demo: Handling resolved threads correctly
We need to be able to resolve threads and have the expected behavior of the thread being removed from the inline threads. Similarly, when re-adding it, we want it to appear.

Finally, when clicking on the Thread from the ThreadedComments, we want to focus on the correct inline thread, however when it's resolved, we just want to focus the thread on ThreadedComments, with no additional action (currently the app explodes when we can't find the clicked thread inside the open threads!

You may notice that I have split the `handleRemoveThread` function into `handleRemoveThread` and `handleHideThread`. I wanted to separate these two behaviors because we actually want to remove threads that we never populate, but just want to hide resolved threads.

Note: I recommend hiding whitespace on the diff

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/456d1217-a997-4ada-ba22-16290565ad3d





Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6808

monorepo-commit: 6bb16ff87a35ab2bda002607078a087d0560c776