New document demo: Thread sorting edge case
When having highlights on the same line of the document demo app, we ended up with kind of random
sorting between the threads in the same line, as their distance from the top was the same. This PR checks
for possible ties in the sorting and breaks the tie by comparing the distance from the left side of the
document.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/531ab74a-c809-4d24-985f-cac44d0c9dfc

Multiple line test plan:

https://github.com/getcord/monorepo/assets/102963329/48d74ef8-9a42-4776-8a97-014bbe4da500





Reviewers: flooey, dmmiller, AlbertQM

Reviewed By: dmmiller, AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6809

monorepo-commit: f83335e16e3d1e2797b62e68c5221b262b607efb