Unhiding seen-by in the new document demo
The seen by indicators were initially hidden in the document demo, to try and make it look as
close as possible to Google Docs. However, this is an important feature demo-ed in our customer calls.
So we decided to show it again.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/67306695-4cda-452a-a9fa-845e369dbdc8




Reviewers: AlbertQM, flooey, elenaperezrioja

Reviewed By: elenaperezrioja, AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6882

monorepo-commit: 0b20107c32642d3203933d48f77cf4875a6a4bf1