Modifying message structure in the new document demo ThreadedComments
When having long messages the fact that the text goes all the way
to the end of the message does not look very good in ThreadedComments in
the new document demo.

I have modified the grid layout of the message component to not extend the
messageContent all the way to the right edge of the message.

Test Plan:
Before:
<img width="1032" alt="Screenshot 2023-10-17 at 15 40 45" src="https://github.com/getcord/monorepo/assets/102963329/e5755c16-8efa-4466-aa2c-5c4d6aa9763a">


After:
<img width="446" alt="Screenshot 2023-10-17 at 15 44 41" src="https://github.com/getcord/monorepo/assets/102963329/5e2840e1-35e7-4b3f-807c-30f650178119">
<img width="523" alt="Screenshot 2023-10-17 at 15 44 49" src="https://github.com/getcord/monorepo/assets/102963329/7b049e52-2d86-4c39-afd3-f7d96754a2ab">


Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6892

monorepo-commit: 37e24056b17972c0b5f1516b1fe67d2ff957a5bf