Demo app dashboard - show resolved threads (#6921)
## Summary

Right now, when a user resolves a thread, nothing really happens to the
icon, so it almost could feel like nothing happened. Then, when the user
refreshes, the icon disappears.

We could have somewhere a switch to show all resolved too, and we might
add that, but for now, we need to at least show the icon a little bit
different when a thread has been marked as resolved at that moment.

There are two keys for this change:

- We were not fetching the resolved threads. When a thread is resolved,
we won't fetch it again, so we don't know it's resolved. But then we
need to fetch ALL resolved, and we don't want to show all of them.
Therefore, we update a thread to resolve only if it was not resolved at
some point before in this session.
- The svg. I could use currentColor. Yes, I know. But then, we cannot
use it as a source im <img /> and we need to import it. I fought a bit
with Vite and couldn't get it to work. I'm sure it's a little thing I'm
missing, but I didn't want to waste more time so I copy pasted the
asset...#notProud

Warning: I'm no design expert, I chose purple, it could be aaaaany other
color.

## Test plan



https://github.com/getcord/monorepo/assets/102948032/8498d0d1-dce5-4869-ae14-d18c6c4a2dc6

monorepo-commit: ebf3fb0b1c0236819c99bce211937ed1a4048057