demo-apps: video player - fix bug of reactions & option menu breaking comment style
There is this awful bug where if you try to react or open the options menu in a comment towards the bottom of the video player, it will push the comment outside of it's container and the options menu and emoji picker will be "stuck" inside of the comment box and be cut off.

Turns out @lazybean is a genius! He helped me find this 🧙‍♂️ The issue was that a class `thread-on-top` adds a `transform` property which creates a new stacking context, so instead of the options menu & emoji picker being fixed positioned in relation to the viewport, it did so in relation to the thread container.

Refactoring this css to not use transform has fixed the issue! Thank you @lazybean!

**Before**

https://github.com/getcord/monorepo/assets/62358728/f1e49f8c-9706-4131-bc11-760fb2d361b4

**After**

https://github.com/getcord/monorepo/assets/62358728/a5a28c02-e80c-47e7-8d4f-00140b3b9081




Test Plan:
Create a comment at the bottom of the video, try to react to the comment or open the options menu. It should open as expected

Reviewers: flooey, #infra, lazybean, elenaperezrioja

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/6978

monorepo-commit: d75e8c3eb42812dc0c67e11baf9f0f24a1499701