New document demo: Hiding comments behind document header
Making the floating threads go behind the header was not a straight forward process, due to only having a two layer stacking context, which we had to change. This PR does a bunch of different things that had to be done together in order for this to work. Specifically:

1. Brings the text highlights and floating threads within the doc container
2. Sets position: relative (from position: static) to the header container so that we can correctly create a stacking context
3. Shuffles some CSS (mostly margins/paddings) around to preserve the layout after changing the positioning of the elements
4. Includes top and left offsets in the calculations of the text highlight and floating thread positioning as they are now rendered relative to their parent, rather than the viewport.

Note 1: We need the bounding client rect of the parent of our container, so we can understand what the offsets for top and left for the text highlights and floating threads are at any moment. However, we don't really need to recalculate it every time as it's expensive and unnecessary. We only need to recalculate when the viewport changes. This would take some amount of code though. The compromise I made was putting the result of the calculation in a ref, and recalculating that value every time we repaint floating threads (instead of before repainting every text highlight, floating threads and floating presence). Then passed the offset coordinates to the components that need them.

Note 2: I recommend using `hide whitespace` on the diff.

Test Plan:
https://github.com/getcord/monorepo/assets/102963329/96b47d68-0cab-445a-a67b-76c053d93890




Reviewers: AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6846

monorepo-commit: a7a035c34e417c5a93ace4e52c39fed68176bdc3