Renaming the --comments-z-index css variable
In the new document demo, we had a variable called `--comments-z-index`. This was a confusing name
as it could be mistaken for the z-index of the floating threads, on the right hand side of the document. It
actually refers to the ThreadedComments component. I have renamed this variable to `--threaded-comments-z-index`
to reflect this more accurately.

Test Plan:
New document demo ThreadedComments have their z-index set to the correct value and the CSS variable is
seen in dev tools.

<img width="246" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/0e33070a-cbd2-47d5-bc25-bdc8ec881a9a">


Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/6922

monorepo-commit: e4ee74c1c1a625663d595414a8f3eab84062ace1