New document demo: Hiding select-emoji button on unfocused thread when we have an open one
When having an open thread and we click to react to a message of an unfocused one,
the emoji picker appeared behind the open thread (I know this is a bit confusing, check
the "before" video attached).

Due to the stacking contexts being really complex and us using transformations in css, which
contributes to the stacking context complexity, we decided that the solution to go for right
now is to just hide the select-emoji button for unfocused floating threads.

An interesting thing is that we only want to hide the button when there's an open (focused)
thread. Otherwise, it is still useful to be able to react to an unforcused floating thread.
For this reason, I've added a classname to the parent container of all floating threads to
indicate whether at least one thread is open. Then, added a classname to the specific open
thread. This works nicely, though I'm not sure is worth the complexity? Please let me know
if this is too much and I should just disable reactions for all unfocused floating threads.

Test Plan:
Before:

https://github.com/getcord/monorepo/assets/102963329/06cd352e-93b5-4473-9037-2c360da60bb5

After:

https://github.com/getcord/monorepo/assets/102963329/71ac153b-2288-4c07-b8f8-5df03e1da580




Reviewers: AlbertQM, lazybean, flooey

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/6890

monorepo-commit: f581384e58b56400ab1bd3c62fc38c3feb131265