Demo apps- document - Add comment bubble scrolls (#7035)
## Summary

When selecting text, the "Add comment" bubble appears. If the user then
scrolls, the bubble keeps its position, which is not close to the
selected text anymore.

We're not using any special libraries, and I'm not going to add one
(floating-ui) just for this. So we just need to save the scroll position
at text selection time, get the current scroll position whenever we
scroll and recalculate the bubble position when any of these change.

## Test plan



https://github.com/getcord/monorepo/assets/102948032/84deaac6-1154-4e1e-88b0-187bbacc766d

monorepo-commit: 02aea7765c921758d6c68ef7a35d048445ff6483