New canvas zoom button (#7072)
## Summary

I had countless problems with @mhtien's branch
(https://github.com/getcord/monorepo/pull/6754). My fault and git's
fault, the code was ok.

So I patched the code and modified some things:

- We should have a centralized place that reflects the current scale. We
already have the scale in the Stage, but it is a ref, and we're not
listening to changes. I added a new state in the context so that we have
the scale there.
- Also moved the function to zoom to the context. It will take the new
scale and an optional center (used when pinching and zooming) and
removes the code duplication that @mhtien's PR had.

I think it'd make sense to have a different number jump when clicking
the + and -, not just a 3%, but we can do that in another PR

## Test plan



https://github.com/getcord/monorepo/assets/102948032/4e72a7d1-8380-45d3-8e56-bbbd7202a560

monorepo-commit: 6572fd09a504699089a7027e8c13937b25804850