Remove thread from canvas (#7111)
## Summary

In the new canvas app, when a thread was deleted, we would still show it
in the canvas. This alone is weird, as the pin should dissappear. But
also, the data rendered inside was all kinds of wonky.

I think we should stop showing the pin when the thread has been deleted.

I also updated the JS call to use non deprecated functions

## Test plan



https://github.com/getcord/monorepo/assets/102948032/a7299ded-3ecf-4f74-898f-3665a4412c8b

monorepo-commit: 68a224e4c013679b4bf41ea99ca3d739faf454fa