Demo video app scroll fix (#7125)
## Summary

The video app is scrolling the whole window whem focusing a message in
the threaded comments component. If the message was clicked is because
it was already visible, there is no need to scroll the window.

Also, when clicking on the message, it depends whether we click on the
"reply" or not, the message will forcus or not. If we click on reply, we
should focus on that message too.

## Test plan



https://github.com/getcord/monorepo/assets/102948032/8becd900-09e8-474c-bee7-67ec5f32b984

monorepo-commit: 5ff75018b74b1eab570e89ff2e4b5aaea44a7874