Document demo: Adding a box-shadow to the ThreadedComments
I noticed that the ThreadedComments-like component in GoogleDocs had a box-shadow, which
made it separate in a clearer way from the document below it. I have added the same box-shadow
to the ThreadedComments component in the document demo.

Test Plan:
The box-shadow correctly appears on the component.

<img width="245" alt="image" src="https://github.com/getcord/monorepo/assets/102963329/75eaa2fb-e3fa-4715-b531-fcbc6bac5207">



Reviewers: elenaperezrioja, flooey, maylynn-ng, km-nur

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/7202

monorepo-commit: 559947dcd7dc7add0f7f519099a5fbe1fbc25cb4