Fix dashboard invisible threads (#7166)
## Summary

Whule working on the "mark notif as seen" feature, I realized that we
are marking as seen threads that are not really being shown. This is
because instead of not rendering them, we are rendering but hiding them.

This causes the message and thread seen observers to be fired for every
message in the charts, even when we're not really showing them.

## Test plan

Load the dashboard and check on the graphql requests that we're not
calling "markThreadSeen" mutation, and that it's only called once the
thread is visible for a few seconds.

monorepo-commit: 8565e2dd6d95ec348bbfd9ab23858c85d09ece13