Document demo: Disabling the selection of the avatar
We found a bug in the document demo where the Avatar indicating presence, on the left side of each paragraph could be highlighted. This caused the "Add comment" button to be positioned incorrectly, as well as the new floating thread on the right side of the document to appear in a completely wrong position (outside the viewport) - video below.

This PR fixes this bug by disabling the selection of the cord-avatar element via CSS.

Test Plan:
Video of the bug:

https://github.com/getcord/monorepo/assets/102963329/c5e2412f-2663-4c4e-bee1-e29ad5f129a4

Test case: Attempting to highlight the avatar, in the same way as above. Then add a comment.
Expected outcome: It's not possible to highlight. The "Add comment" button and floating thread appear in their respective correct positions.

Test case: Select the PresenceFacepile on the top right of the doc, the Avatars inside the message component and the Avatars inside the ThreadedComments component.
Expected outcome: We should be able to highlight all of the aforementioned components.

https://github.com/getcord/monorepo/assets/102963329/82b99bef-f21d-4efd-ab44-50d159fb5bdd







Reviewers: maylynn-ng, flooey

Reviewed By: maylynn-ng

Pull Request: https://github.com/getcord/monorepo/pull/7184

monorepo-commit: 288107f32cf8e91fcc643a657c2b698593184b3d