Canvas demo: Adding avatars to grouped pins
The canvas demo currently groups pins together when the distance between the pins is under 50px. When this grouping happens, the pins are replaced with a white bubble, which indicates the number of threads in that group. We want to change this behavior to match exactly that of Figma, which initially shows the avatars of the thread authors within the white bubble, and on hover, shows the number of threads in the group.

Since the solution required hardcoding styles for each grouping number, we decided to limit the grouping to show at maximum 3 avatars in a bubble. The number was chosen based on the 2 authors of the dummy comments we have by default in the docs sample apps, plus the viewer of the docs.

Test Plan:
The initial state is completely zoomed in, so no groups exist. There are 6 total threads by 3 different authors.

Test case: Zoom out, comments by the same author are grouped.
Expected outcome: The group has two avatars, the avatar of the author and right behind it, the same avatar, but 50% faded ("ghost" avatar). On hover, the white bubble shows the number "2".

Test case: Zoom out more, another comment from the same author gets added in the same group.
Expected outcome: The same avatars as the previous case appear, but on hover, the number "3" appears in the bubble.

Test case: Zoom out more, comments by different authors are grouped
Expected outcome: The bubble of the new group shows a layout with 2 avatars and the number "2" on hover.

Test case: Zoom out completely, until all pins are grouped
Expected outcome: The bubble shows the layout of 3 avatars and the number "6" on hover - for the total threads.


https://github.com/getcord/monorepo/assets/102963329/647ba69a-a230-412e-a588-19be0fb74979





Reviewers: flooey, maylynn-ng, elenaperezrioja, km-nur

Reviewed By: elenaperezrioja

Pull Request: https://github.com/getcord/monorepo/pull/7212

monorepo-commit: b16b45d90913fc87dfe16843baafe5004a841d4d